backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
google-genai
SpeechRecognition
pyaudio
pydub
gTTS
pytest
pydantic
